requirements.txt
openai
python-dotenv
pydantic==1.*
langchain
langchain_openai
langchain_community
streamlit